+++ lists/src/test/kotlin/mathtools/lists/doubles/DoubleListGreaterThanTest.kt
@@ -1,17 +1,17 @@
-package mathtools.numbers.listtypes.doubles
+package mathtools.lists.doubles
 
 import mathtools.lists.DoubleList
-import mathtools.numbers.listtypes.doubles.DoubleListTestResources.largeList
-import mathtools.numbers.listtypes.doubles.DoubleListTestResources.largeListFactor
-import mathtools.numbers.listtypes.doubles.DoubleListTestResources.maxValueList
-import mathtools.numbers.listtypes.doubles.DoubleListTestResources.medList
-import mathtools.numbers.listtypes.doubles.DoubleListTestResources.medListFactor
-import mathtools.numbers.listtypes.doubles.DoubleListTestResources.nanValueList
-import mathtools.numbers.listtypes.doubles.DoubleListTestResources.negativeInfiniteList
-import mathtools.numbers.listtypes.doubles.DoubleListTestResources.positiveInfiniteList
-import mathtools.numbers.listtypes.doubles.DoubleListTestResources.singleItemList
-import mathtools.numbers.listtypes.doubles.DoubleListTestResources.smallList
-import mathtools.numbers.listtypes.doubles.DoubleListTestResources.smallListFactor
+import mathtools.lists.doubles.DoubleListTestResources.largeList
+import mathtools.lists.doubles.DoubleListTestResources.largeListFactor
+import mathtools.lists.doubles.DoubleListTestResources.maxValueList
+import mathtools.lists.doubles.DoubleListTestResources.medList
+import mathtools.lists.doubles.DoubleListTestResources.medListFactor
+import mathtools.lists.doubles.DoubleListTestResources.nanValueList
+import mathtools.lists.doubles.DoubleListTestResources.negativeInfiniteList
+import mathtools.lists.doubles.DoubleListTestResources.positiveInfiniteList
+import mathtools.lists.doubles.DoubleListTestResources.singleItemList
+import mathtools.lists.doubles.DoubleListTestResources.smallList
+import mathtools.lists.doubles.DoubleListTestResources.smallListFactor
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.RepeatedTest
 import org.junit.jupiter.api.Test
